pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "disruptsc"
dynamic = ["version"]
description = "A spatial agent-based model to simulate the dynamics supply chains subject to disruptions"
readme = "README.md"
license = {text = "CC BY-NC-ND 4.0"}
authors = [
    {name = "Celian Colon", email = "celian.colon.2007@polytechnique.org"}
]
maintainers = [
    {name = "Celian Colon", email = "celian.colon.2007@polytechnique.org"}
]
keywords = ["supply-chain", "agent-based-model", "disruption", "spatial", "economics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: GIS",
]
requires-python = ">=3.10,<3.12"
dependencies = [
    # Core data processing
    "pandas>=1.3.0,<3.0",
    "numpy>=1.20.0,<2.0",
    
    # Geospatial processing
    "geopandas>=0.11.1,<1.0",
    "shapely>=1.8.0,<3.0",
    
    # Network analysis
    "networkx>=2.8.0,<4.0",
    
    # Scientific computing
    "scipy>=1.7.0,<2.0",
    
    # Configuration and utilities
    "PyYAML>=5.3.0,<7.0",
    "tqdm>=4.60.0,<5.0"
]

[project.urls]
Homepage = "https://github.com/ccolon/disruptsc"
Documentation = "https://ccolon.github.io/disruptsc"
Repository = "https://github.com/ccolon/disruptsc.git"
Issues = "https://github.com/ccolon/disruptsc/issues"

[project.scripts]
disruptsc = "disruptsc.main:main"
validate-inputs = "disruptsc.model.input_validation:main"

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
    "pre-commit",
]
docs = [
    "mkdocs",
    "mkdocs-material",
    "mkdocstrings[python]",
]
jupyter = [
    "jupyter",
    "matplotlib",
    "plotly",
    "ipywidgets",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "disruptsc._version.__version__"}

[tool.black]
line-length = 120
target-version = ['py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/disruptsc --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "geopandas.*",
    "shapely.*",
    "networkx.*",
    "scipy.*",
    "tqdm.*",
]
ignore_missing_imports = true